Add files via upload
I noted this a few times in the code, but there was an issue with my goes east/west transformations. I couldn't figure out why it was like that but attempted to get around it in a few ways. Ughhhhh at least my computer remembered how to find a module after a week or so...